pyproject.toml
[project]
name = "legal-translator"
version = "0.1.0"
description = "Minimal viable prototype for batch legal document translation with Claude."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [
  { name = "Translator MVP Team" }
]
dependencies = [
  "anthropic>=0.35.0",
  "python-docx>=1.1.0",
  "pdfplumber>=0.11.4",
  "PyPDF2>=3.0.1",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "python-multipart>=0.0.9",
  "rapidfuzz>=3.10.0",
  "python-dotenv>=1.0.1",
  "typer>=0.12.5",
  "pydantic>=2.9.2",
  "tenacity>=9.0.0",
  "reportlab>=4.0.0",
  "ddgs>=9.0.0",
  "beautifulsoup4>=4.12.0",
  "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "ruff>=0.7.4",
  "mypy>=1.13.0"
]

[project.scripts]
translate = "translator.cli:app"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"


